Allow users to specify data interface along with mgmt interface
Currently there is only interface - mgmt interface specified in node configuration.
Add another interface - data interface - that may be used for drivers as they wish.
The mgmt interface is used for mgmt purposes and gossip.
The data interface, if not specified, defaults to mgmt interface.
If data interface is specified but mgmt interface is not, mgmt interface defaults to data interface.
If both are not specified, the first available interface is used for both.