Initialize all drivers with snapshot of the init kv store.
In a clustered setup, the central kv store may keep changing all the time,
so the state at which the cluster manager starts and creates init database
 may become inconsistent by the time other components start and refer to
this init database. Therefore create a snapshot of the kv store on startup
and pass it init/join phase so that all the components can start with the
same state of the kv store.