Add forceRemove flag for node decommission.
If forceRemove is true, all cluster listeners must do whatever
required to remove the node or mark it decommissioned.
Some sanity checks still return an error, for instance, even if force is
set to true currently we do not allow decommissioning a node which is
not offline or in maintenance mode (may change in future).