Support for non-quorum members.
This diff introduces support for nodes to be defined as non-quorum
members. Such members do not participate in quorum decisions.
An example of such nodes is pure compute-only nodes. Since they do not
have any state they do not need to participate in quorum-decisions.

The cluster manager queries listeners if they want to participate in
quorum decision - if all listeners reply in negative, the node becomes a
non-quorum node.

The cluster remains in quorum as long as majority of nodes which are
quorum deciding nodes remain operational.