Reorganize to create cluster.ClusterClient
This change also makes the struct api/client/cluster.ClusterClient
exposed and moves the documentation there. We will be adding examples
and more documentation as time continues.

This reoganization of the interfaces and structs makes it simple for
developers to read the pages generated by GoDoc. When we add examples,
those examples could then be run right from the webpage using
play.golang.org as it is done in the official Golang packages
documentation.

Closes #382

Signed-off-by: Luis Pabón <luis@portworx.com>