Add in-memory schedule policy to cluster
The issue is that even though it compiled, the current cluster.Cluster
instance in osd did not have an implementation of the schedulerpolicy.
This would cause panics when the SDK would try to use the functions.
Instead it was in the REST server.

This patch provides the following:

* Move the configuration struct to cluster/ from api/server
* Add an in-memory implementation of the SchedulePolicy implementation
* Change the interface name from SchedulePolicy to
SchedulePolicyProvider so that it does not conflict with the struct in
schedpolicy/ of the same name.

Closes #487

Signed-off-by: Luis Pabón <lpabon@gmail.com>