Add a new NULL implementation for Cluster interface.
The cluster pkg has two primary interfaces
- Cluster
- ClusterListener

The ClusterManger in openstorage implements the Cluster interface. It also invokes the APIs
from the ClusterListener interface. Whoever implements the ClusterListener interface will get
events from the ClusterManager. By adding a NULL implementation we would be able to write tests
for the implementors for ClusterListener interface.

Signed-off-by: Aditya Dani <aditya@portworx.com>