Add new Uuid() API to Cluster
Cloud and Backup APIs require a unique identifier and not the
Cluster.Id. Therefore a new interface was added to retrieve the
unique id of the cluster.

To standardize objects in the SDK, the old cluster id has been
moved to the StorageCluster.Name since, like a Volume object, it
has been provided by the caller or admin. The StorageCluster.Id
now will be filled in with the unique identifier if available.

Signed-off-by: Luis Pabón <luis@portworx.com>