Modify DeviceMount mounter type to handle device symbolic links.
A lot of devices passed to mounter could be symbolic links pointing to a different device.
For ex. in case of LVM the device passed to mounter could be /dev/mapper/lvm1 which has a symbolic link to /dev/dm-0
The underlying mount table can inter changeably use either of the device which might cause mounter to return
a device as not mounted even if it is mounted but with a different device name.

This change handles such devices by adding an entry to the mounter's mount table for the target device in the Load API.

Signed-off-by: Aditya Dani <aditya@portworx.com>