Add support for forcing fs to spec_handler.go
This change is needed to allow the value to be set from
a StorageClass in K8S. Forcing the unsupported fs type
will allow callers to ask the driver to use an unsupported
fs type if possible. For example, using the Portworx driver,
the administrator can force XFS as the file system of the
volume if this setting is set to true.

Closes #633

Signed-off-by: Luis Pabón <lpabon@gmail.com>