Don't print HTTP status in error message
There is StatusCode() if callers want that. This causes HTTP errors to be
propagated up to the APIs and printed for users which isn't relavent. Also for
an API making other API calls which failed this would print multiple HTTP error
codes.

Also trim newlines from the end of the body. The APIs we use add extra newlines
at the end

Signed-off-by: Dinesh Israni <disrani@portworx.com>