NFS Raw block support
Volume Creation:
Will attach to the NFS host, then create a file of the requested
size on the NFS share. It will then use losetup to mount the file
and format it according to the format in the request. If not format
was requested, then it will be left as RAW.

Snapshots are done by copying the file.

THIS IS NOT FOR PRODUCTION, ONLY FOR TESTING of Raw block store

Signed-off-by: Luis Pabón <luis@portworx.com>