sdk: Add options to mountattach service RPCs
When we created the MountAttach service for the SDK, we removed support for
map[string]string because it was confusing for developers to use since the
values are not very clear. What we didn't realize is that drivers may use
this payload to add unique, opaque information.

This patch adds the option for driver specific values to be passed
in during the use of MountAttach service calls.

Signed-off-by: Luis Pabón <luis@portworx.com>