Support for multitenancy in intree Kubernetes
This change fixes a security issue where the pvc user could
maliciously use the secret located in a namespace they cannot
access. Instead this patch changes the model to the CSI model
which is based on using the StorageClass as the source of truth
on the location of the sercret which contains the authentication
token.

Signed-off-by: Luis Pabón <luis@portworx.com>