Fix token in middleware server
When a request with a token and a user agent was sent to the
old REST server, the auth middleware assumed that it was from
Kubernetes and tried to get the token for the request, ignoring
the token provided.

This fix makes sure that the token provided is used instead.

Signed-off-by: Luis Pabón <luis@portworx.com>