Make unit tests compatible with podman
Docker is starting to be replaced by podman, rc, and others. This
change makes it possible to run the unit tests using podman on
the development system. There seems to be an issue with mouting
inside podman and using chattr. By default the unit tests will
not be changed.

A developer using podman or another container runtime can use:

    make TAGS=daemon docker-test

Signed-off-by: Luis Pabón <lpabon@purestorage.com>