PWX-21380: return a deep copy of the peer api.Node to the caller (#1946)
We were deep-copying the api.Node struct for the self node but not for the
peer node. Moreover, the peer node struct is a shallow copy of the one
inside the gossip store. Any modifications to this struct can trigger "concurrent
map read and map write" panic when gossip is trying to send this struct out.

This interim fix avoids the issue by returning a deep copy of the peer
api node struct. There is still more work left to ensure that the gossip
store does not return a shallow copy of its api.Node struct.

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>