PWX-26578: fix the slow task scheduling when runOnce tasks are present (#2103)
In the scheduleTasks function, we iterate over all the tasks to find the ready tasks and
feed them to the workers. When we encounter a runOnce task, we feed it to the workers and
then remove it from the task list. But the removal resets the current element's next pointer
to nil. This terminates the iteration prematurely causing other ready tasks to not get added
to the work queue.

Now, we remove the runOnce tasks outside of the iteration.

Also, added dynamic creation of new worker goroutines.

- start a new batch (10) of worker goroutines if the number of enqueued tasks
  is high enough

- excess workers exit after being idle for 30 seconds

- total number of workers is capped at 100

Also,

- fixed staticcheck warnings

- added some basic stats collection to the scheduler

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>