PWX-25794: add enum value for not defaulting to sharedv4 service (#2117)
When we start defaulting to "sharedv4 service" volumes, we need
a way to create sharedv4 volumes. This patch adds a new value
NONE to the service type enum. This new value can be used
to indicate that the caller does not want "sharedv4 service" volume.
The new enum value is mapped to an empty string in the spec to be compatible
with the value used when converting sharedv4 svc volume to sharedv4
volume.

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>