Fix osd-dev test pt1 - skip comparing protobuf parameters
OSD-dev was failing for a few reason. First the test is expecting to use Go
1.16, so we updated the osd-dev image to use 1.16. Secondly, there is issue
with mock comparing structs generated by protobuf. The new protobuf version
has extra hidden fields that causes comparison to fail even if the values
are the same. Similar issues in https://github.com/stretchr/testify/issues/758.

For the fix, we move all of the mock request using protobuf to use
mock.Any() as the parameters to bypass the comparison. This should be
discussed with other contributors if there are other possibilities.

Testing: local testing passes this issue
Signed-off-by: dahuang <dahuang@purestorage.com>