PWX-36511 Expose API to return a set of nodes that can be upgraded in parallel (#2440)
* PWX-36511 Expose API to return a set of nodes that can be upgraded

in parallel. This takes an input list of nodes from which the server will return
a set that doesn't have any common volume within them. This ensures that if
these nodes are upgraded at once, then a volume with 1+ replica has at least one
replica somewhere which will be up.

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>

* PWX-36511 Add interface to volume driver to support upgrade funcs

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>

---------

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>
Co-authored-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>